Repackage Java classes in accordance with the Bounded Context nesting.
