+++ server/src/main/java/io/spine/quickstart/tasks/server/TasksContext.java
@@ -18,10 +18,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.quickstart;
+package io.spine.quickstart.tasks.server;
 
-import io.spine.quickstart.server.QuickStartStorageFactory;
-import io.spine.quickstart.task.TaskAggregate;
+import io.spine.quickstart.tasks.server.task.TaskAggregate;
 import io.spine.server.BoundedContext;
 import io.spine.server.CommandService;
 import io.spine.server.DefaultRepository;
@@ -38,7 +37,7 @@ public final class TasksContext {
     /**
      * The name of the context.
      */
-    public static final String NAME = "Tasks";
+    static final String NAME = "Tasks";
 
     private static final BoundedContext context = createContext();
