+++ server/src/test/java/io/spine/quickstart/tasks/server/task/TaskAggregateTest.java
@@ -18,12 +18,12 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.quickstart.task;
+package io.spine.quickstart.tasks.server.task;
 
-import io.spine.quickstart.tasks.command.CreateTask;
-import io.spine.quickstart.tasks.task.Task;
-import io.spine.quickstart.tasks.event.TaskCreated;
 import io.spine.quickstart.tasks.TaskId;
+import io.spine.quickstart.tasks.command.CreateTask;
+import io.spine.quickstart.tasks.event.TaskCreated;
+import io.spine.quickstart.tasks.task.Task;
 import io.spine.server.DefaultRepository;
 import io.spine.testing.server.blackbox.BlackBoxBoundedContext;
 import org.junit.jupiter.api.DisplayName;
