+++ web/server/src/main/java/io/spine/quickstart/tasks/web/TasksSubscribeServlet.java
@@ -18,27 +18,27 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.quickstart.web;
+package io.spine.quickstart.tasks.web;
 
-import io.spine.web.query.QueryServlet;
+import io.spine.web.subscription.servlet.SubscribeServlet;
 
 import javax.servlet.annotation.WebServlet;
 
 /**
- * {@code Tasks} context query servlet.
+ * {@code Tasks} context {@code /subscribe} servlet.
  *
- * <p>Handles the queries {@code POST}ed by the client by dispatching them to
- * the {@link io.spine.web.firebase.query.FirebaseQueryBridge}.
+ * <p>This is a part of the system's subscription web API. Handles the subscriptions created by
+ * the client via the {@link io.spine.web.firebase.subscription.FirebaseSubscriptionBridge}.
  *
- * @see QueryServlet
- * @see io.spine.quickstart.web.ServletBridges
+ * @see SubscribeServlet
+ * @see ServletBridges
  */
-@WebServlet("/query")
-public final class TasksQueryServlet extends QueryServlet {
+@WebServlet("/subscription/create")
+public final class TasksSubscribeServlet extends SubscribeServlet {
 
     private static final long serialVersionUID = 0L;
 
-    public TasksQueryServlet() {
-        super(ServletBridges.query());
+    public TasksSubscribeServlet() {
+        super(ServletBridges.subscription());
     }
 }
